created hook for functionality
